submissionscript/requirements.txt
astroid==2.15.4
black==23.3.0
click==8.1.3
dill==0.3.6
flake8==6.0.0
flake8-black==0.3.6
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
protobuf==3.20.3
pycodestyle==2.10.0
pyflakes==3.0.1
pylint==2.17.4
tomlkit==0.11.8
wrapt==1.15.0
